Adds SopService

+++ backend/src/main/java/com/woven/app/repository/SopRepository.java
@@ -3,5 +3,5 @@ package com.woven.app.repository;
 import com.woven.app.domain.Sop;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SopRepository extends JpaRepository<Sop, Long> {
+public interface SopRepository extends JpaRepository<Sop, Integer> {
 }
