+++ backend/src/main/java/com/woven/app/repository/ToolRepository.java
@@ -3,5 +3,5 @@ package com.woven.app.repository;
 import com.woven.app.domain.Tool;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ToolRepository extends JpaRepository<Tool, Long> {
+public interface ToolRepository extends JpaRepository<Tool, Integer> {
 }
